add Product domain

+++ app/src/main/java/com/codesoom/assignment/domain/Product.java
@@ -16,7 +16,7 @@ public class Product {
 
     private String name;
 
-    private Long price;
+    private int price;
 
     public Long getId() {
         return id;
@@ -34,11 +34,11 @@ public class Product {
         this.name = name;
     }
 
-    public Long getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(Long price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
